Safety checkin - transfer

+++ include/Transmitter.h
@@ -28,25 +28,76 @@
 #include "FileDeliveryTable.h"
 
 namespace LibFlute {
+  /**
+   *  FLUTE transmitter class. Construct an instance of this to send data through a FLUTE/ALC session.
+   */
   class Transmitter {
     public:
+     /**
+      *  Definition of a file transmission completion callback function that can be
+      *  registered through ::register_completion_callback.
+      *
+      *  @param toi TOI of the file that has completed transmission
+      */
       typedef std::function<void(uint32_t)> completion_callback_t;
 
+     /**
+      *  Default constructor.
+      *
+      *  @param address Target multicast address
+      *  @param port Target port 
+      *  @param tsi TSI value for the session 
+      *  @param mtu Path MTU to size FLUTE packets for 
+      *  @param io_service Boost io_service to run the socket operations in (must be provided by the caller)
+      */
       Transmitter( const std::string& address, 
           short port, uint64_t tsi, unsigned short mtu,
           boost::asio::io_service& io_service);
 
+     /**
+      *  Default destructor.
+      */
       virtual ~Transmitter();
 
+     /**
+      *  Enable IPSEC ESP encryption of FLUTE payloads.
+      *
+      *  @param spi Security Parameter Index value to use
+      *  @param key AES key as a hex string (without leading 0x). Must be an even number of characters long.
+      */
       void enable_ipsec( uint32_t spi, const std::string& aes_key);
 
-      uint64_t send(const std::string& content_location,
+     /**
+      *  Transmit a file. 
+      *  The caller must ensure the data buffer passed here remains valid until the completion callback 
+      *  for this file is called.
+      *
+      *  @param content_location URI to set in the content location field of the generated FDT entry
+      *  @param content_type MIME type to set in the content type field of the generated FDT entry
+      *  @param expires Expiry timestamp (based on NTP epoch)
+      *  @param data Pointer to the data buffer (managed by caller)
+      *  @param length Length of the data buffer (in bytes)
+      *
+      *  @return TOI of the file
+      */
+      uint16_t send(const std::string& content_location,
           const std::string& content_type,
           uint32_t expires,
           char* data,
           size_t length);
+
+     /**
+      *  Convenience function to get the current timestamp for expiry calculation
+      *
+      *  @return seconds since the NTP epoch
+      */
       uint64_t seconds_since_epoch();
 
+     /**
+      *  Register a callback for file transmission completion notifications
+      *
+      *  @param cb Function to call on file completion
+      */
       void register_completion_callback(completion_callback_t cb) { _completion_cb = cb; };
 
     private:
